+++ src/core/pipe/units/column.pipe.unit.js
@@ -20,16 +20,17 @@ export const columnPipeUnit = [
 			behavior: 'core'
 		};
 
-		const columns = columnLine.map(c => c.model);
-		model.view({ columns }, tag);
+		model.view({
+			columns: columnLine.map(c => c.model)
+		}, tag);
 
-		const column = {
-			rows: scene.columnRows(memo.columns),
-			area: scene.columnArea(memo.columns),
-			line: columnLine
-		};
-
-		context.model.scene({ column }, tag);
+		context.model.scene({
+			column: {
+				rows: scene.columnRows(memo.columns),
+				area: scene.columnArea(memo.columns),
+				line: columnLine
+			}
+		}, tag);
 
 		next(memo);
 	}